Modloader: Fix bug where ELF isn't patched/unpatched
Due to one side of a string comparison I do being left case sensitive.  (when looking for the path in the ELF to patch.

Caused it to not finding the path anymore when disabling a mod, if the p2m mod has an int/olm/wp2/e.g with a lowercase file name.