Modloading: Improved Performance
NEW P2M VERSION - 4, no support for v3 or lower.
This seemingly eliminates slowdown that was noticed with large mods - tested on one with 1000 files.

fgets calls were becoming a bottleneck, so P2M and activemods.cache formats have been refactored to store strings in more efficient way that doesnt require fgets (length prefixed instead of null terminated.

the modloader has also been refactored to only do file IO with activemods.cache when really necessary, before it was opening the file on every get call, when we should be storing it in memory

Currently modspriority.cache still uses fgets calls but this is fine at the moment as it does not store every mod file. Performance issues would only start to happen if the user is enabling hundreds of mods.