INT Hook performance improvement
While trying to find the root cause of the stuttering - see last commit - I ran with profiling and noticed poor performance when the hook checks if a file is contained in the currently active mods - due to iterating over the whole active mods vector to look for a match. I have improved this by sorting the vector and then running a lower bound binary search instead.
This should be refactored in the future - There are other functions that iterate through the vector that should be replaced, and then I would sort the vector somewhere else. But for now this works to eliminate a performance issue in the INT hook loader.