Fix an issue where overlays weren't rendering
Currently it's possible, for the ui layout update to happen before the game rect has been drawn, leading to overlays being drawn offscreen. This commit adds a check to see if the game rect height/width is 0, and if so, flag for it to try and calculate ui layout again the next frame.